requirements.txt
# Mirror AI - 智能形象进化系统
# Dependencies

# Core ML
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Face Analysis
mediapipe>=0.10.0

# Face Recognition (InsightFace)
insightface>=0.7

# Vector Database
faiss-cpu>=1.7.0

# Scientific Computing
scikit-image>=0.21.0
scipy>=1.11.0

# UI Framework
streamlit>=1.28.0

# Visualization
plotly>=5.18.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
